+++ src/pages/groupComparison/GroupComparisonUtils.ts
@@ -1,10 +1,10 @@
-import {SampleIdentifier} from "../../shared/api/generated/CBioPortalAPI";
+import { SampleIdentifier } from "../../shared/api/generated/CBioPortalAPI";
 import _ from "lodash";
 
 export type SampleGroup = {
-    id:string,
-    name?:string,
-    sampleIdentifiers:SampleIdentifier[],
+    id: string,
+    name?: string,
+    sampleIdentifiers: SampleIdentifier[],
     color?: string;
     legendText?: string;
 };
@@ -28,4 +28,53 @@ export function getCombinations(groups: { name: string, cases: string[] }[]) {
     }
     f({ groups: [], cases: [] }, groups);
     return combinations;
+}
+
+export function getStackedBarData(combinationSets: { groups: string[], cases: string[] }[], categoryToColor: { [cat: string]: string }) {
+    const overlappingCases = _.uniq(_.reduce(combinationSets, (acc, next) => {
+        if (next.groups.length > 1) {
+            acc = acc.concat(next.cases)
+        }
+        return acc;
+    }, [] as string[]))
+
+    let groupedSet = _.reduce(combinationSets, (acc, next) => {
+        if (next.groups.length === 1) {
+            let cases = _.difference(next.cases, overlappingCases)
+            acc[next.groups[0]] = {
+                x: '',
+                y: cases.length,
+                cases: cases,
+                fill: categoryToColor[next.groups[0]],
+                groupName: next.groups[0]
+            }
+        }
+        return acc;
+    }, {} as { [id: string]: { x: string, y: number, cases: string[], fill: string, groupName: string } })
+
+    let groups = _.values(groupedSet).sort((a, b) => a.y - b.y).map(group => [group])
+
+    if (overlappingCases.length > 0) {
+        return [[{
+            x: '',
+            y: overlappingCases.length,
+            cases: overlappingCases,
+            fill: "#CCCCCC",
+            groupName: 'Overlapping Cases'
+        }], ...groups]
+    }
+    return groups
+}
+
+export function getVennPlotData(combinationSets: { groups: string[], cases: string[] }[]) {
+    let maxCount = _.max(combinationSets.map(set => set.cases.length))!;
+    return combinationSets.map(set => {
+        return {
+            count: set.cases.length,
+            //this is to make sure all the circle groups are of same size
+            size: set.groups.length === 1 ? maxCount : set.cases.length,
+            label: `${set.cases.length}`,
+            sets: set.groups
+        }
+    }).sort((a, b) => b.count - a.count);
 }
